fix: Remove cyclic dependency
